+++ api/CHANGELOG.md
@@ -24,6 +24,9 @@
 - Added `Saved` dates to saved patched apps in the `Apps` tab https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/145
 - Added a refresh/reload button to the custom file picker
 - Improved the UI of export and saving dialogs for the custom file picker
+- Updated the view patches screen for patch bundles on the `Discover patch bundles` page to use the same UI as the view patches screen for imported patch bundles
+- Made version tags on patches on all view patch screens searchable with the user set search engine
+- Added patch options/sub-options to the view patches screen on the `Discover patch bundles` page. This is currently only implemented for patch bundles imported from the discovery page as the API dose not currently support patch option fetching for non-imported bundles
 
 
 # Bug fixes